+++ lib/src/pie_theme.dart
@@ -48,6 +48,8 @@ class PieTheme {
     this.angleOffset = 0,
     this.customAngle,
     this.customAngleAnchor = PieAnchor.center,
+    this.menuAlignment,
+    this.menuDisplacement = Offset.zero,
     this.buttonSize = 56,
     this.pointerSize = 40,
     this.tooltipPadding = const EdgeInsets.all(32),
@@ -115,6 +117,18 @@ class PieTheme {
 
   /// Action display alignment for the specified [customAngle].
   final PieAnchor customAngleAnchor;
+
+  /// Alignment of the menu relative to the menu child.
+  ///
+  /// Can be used to display the menu at a specific position
+  /// regardless of the pressed offset.
+  /// For example, you can set it to [Alignment.center] to align
+  /// the menu at the center of the child widget.
+  /// You can combine it with [menuDisplacement] to fine-tune the position.
+  final Alignment? menuAlignment;
+
+  /// Displacement offset for the menu.
+  final Offset menuDisplacement;
 
   /// Size of [PieButton] circle.
   final double buttonSize;
@@ -226,6 +240,8 @@ class PieTheme {
     double? angleOffset,
     double? customAngle,
     PieAnchor? customAngleAnchor,
+    Alignment? menuAlignment,
+    Offset? menuDisplacement,
     double? buttonSize,
     double? pointerSize,
     EdgeInsets? tooltipPadding,
@@ -262,6 +278,8 @@ class PieTheme {
       angleOffset: angleOffset ?? this.angleOffset,
       customAngle: customAngle ?? this.customAngle,
       customAngleAnchor: customAngleAnchor ?? this.customAngleAnchor,
+      menuAlignment: menuAlignment ?? this.menuAlignment,
+      menuDisplacement: menuDisplacement ?? this.menuDisplacement,
       buttonSize: buttonSize ?? this.buttonSize,
       pointerSize: pointerSize ?? this.pointerSize,
       tooltipPadding: tooltipPadding ?? this.tooltipPadding,